test: incluindo teste unitario no projeto fintech

+++ fintech-backend/src/main/java/br/com/fiap/fintech_backend/service/CategoriaService.java
@@ -26,7 +26,11 @@ public class CategoriaService {
         return categoriaRepository.save(categoria);
     }
 
-    public void deleteById(Long id) {
-        categoriaRepository.deleteById(id);
+    public boolean deleteById(Long id) {
+        if (categoriaRepository.existsById(id)) {
+            categoriaRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
